Integrated new process-launching proxy class into server

+++ Broadway.js
@@ -38,36 +38,46 @@ Broadway.prototype.getGtkApplicationExecutableArgs = function getGtkApplicationE
  */
 Broadway.prototype.connect = function connect() {
 	// TODO: Copy parent process env?
-	var broadwayEnv = { 'GDK_BACKEND': 'broadway', 'BROADWAY_PORT': this.getPort() };
+	var broadwayEnv = { 'GDK_BACKEND': 'broadway', 'BROADWAY_PORT': this.getPort() },
+		that = this;
 	
 	console.log('Spawning broadway process');
 	this.process = spawn(this.getGtkApplicationExectutable(), 
 						 this.getGtkApplicationExecutableArgs(), 
-						 broadwayEnv);
+						 { env: broadwayEnv });
 
 	this.process.stderr.on('data', function onBroadwayError(data) {
 		console.log('ERROR from broadway process: ' + data);
-		this.disconnect();
+		that.disconnect();
 	});
 	this.process.stdout.on('data', function onBroadwayOutput(data) {
 		console.log('Broadway stdout: ' + data);
 	});
 
-	console.log('Connecting to broadway via websocket');
-	this.webSocket = new WebSocket('ws://localhost:' + this.getPort() + '/socket', "broadway");
-	this.webSocket.onmessage = function socketDataReceived(event) {
-		this.emit('message', event.data);
-	};
+	setTimeout(function() {
+		var self = that;
+		console.log('Connecting to broadway via websocket: ' + 'ws://localhost:' + self.getPort() + '/socket');
+		self.webSocket = new WebSocket('ws://localhost:' + self.getPort() + '/socket', "broadway");
+		self.webSocket.onmessage = function socketDataReceived(event) {
+			self.emit('message', event.data);
+		};
+	}, 1000);
 };
 
 Broadway.prototype.disconnect = function disconnect() {
 	if (this.webSocket) {
 		this.webSocket.close();
 	}
-	if (this.process) {
+	if (this.process && this.process.killed === false) {
 		this.process.kill('SIGTERM');
 	}
 	this.emit('disconnect');
+}
+
+Broadway.prototype.send = function send(msg) {
+	if (this.webSocket) {
+		this.webSocket.send(msg);
+	}
 }
 
 /*
